#0: fix untilize, reshape, slice, skip gs due to pcc in transpose with untilize
