+++ ttnn/cpp/ttnn/operations/data_movement/slice/device/slice_op.hpp
@@ -26,8 +26,6 @@ struct Slice {
     operation::ProgramWithCallbacks create_program(
         const std::vector<Tensor>& input_tensors, std::vector<Tensor>& output_tensors) const;
 
-    const operation::Hash compute_program_hash(
-        const std::vector<Tensor> &input_tensors) const;
 };
